+++ app/project-structure/shared/text-section.tsx
