+++ app/project-structure/skills/skill-types.ts
@@ -1,0 +1,13 @@
+export enum SkillLevels {
+  Basic = 1,
+  Intermediate,
+  Advanced,
+  Proficient,
+  Master,
+  Expert,
+}
+
+export interface Skill {
+  name: string;
+  level: SkillLevels;
+}
