+++ app/project-structure/skills/skill.tsx
@@ -1,18 +1,5 @@
 import { Progress } from "@/components/ui/progress";
-
-export enum SkillLevels {
-  Basic = 1,
-  Intermediate,
-  Advanced,
-  Proficient,
-  Master,
-  Expert,
-}
-
-export interface Skill {
-  name: string;
-  level: SkillLevels;
-}
+import { Skill, SkillLevels } from "./skill-types";
 
 export const SkillIndicator = ({ level }: Omit<Skill, "name">) => {
   const maxNumberOfLevels = Object.keys(SkillLevels).length / 2; // Since it's a enum, we need to divide by 2 to get the number of levels
