+++ app/project-structure/work-experience/work-experience-types.tsx
@@ -1,0 +1,10 @@
+export interface WorkExperience {
+  company: string;
+  position: string;
+  responsibilities: Responsability[];
+  date: string;
+}
+export interface Responsability {
+  responsibility: string;
+  oneWordSummary: string;
+}
